Implemented the Physical Memory Manager

+++ include/io/serial.h
@@ -15,7 +15,7 @@ namespace io {
     void serialWriteChar(char c);
     void serialWrite(string input);
     void serialWriteHex(u64 num);
-    void serialWriteNumber(u32 num);
+    void serialWriteNumber(u64 num);
 }
 
 #endif //AVERY_SERIAL_H
